+++ source/tests/universal/common/cases/atomic_model/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
