+++ source/tests/universal/common/cases/model/utils.py
@@ -1,5 +1,7 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
 from typing import (
+    Any,
+    Callable,
     List,
 )
 
@@ -9,14 +11,9 @@ from deepmd.dpmodel.utils.nlist import (
     extend_input_and_build_neighbor_list,
 )
 
-from ..utils import (
-    CommonTestCase,
-    forward_wrapper,
-)
 
-
-class AtomicModelTestCase(CommonTestCase):
-    """Common test case for atomic model."""
+class ModelTestCase:
+    """Common test case for model."""
 
     expected_type_map: List[str]
     """Expected type map."""
@@ -34,6 +31,8 @@ class AtomicModelTestCase(CommonTestCase):
     """Expected output type for the model."""
     expected_sel: List[int]
     """Expected number of neighbors."""
+    forward_wrapper: Callable[[Any], Any]
+    """Calss wrapper for forward method."""
 
     def test_get_type_map(self):
         """Test get_type_map."""
@@ -65,6 +64,13 @@ class AtomicModelTestCase(CommonTestCase):
         for module in self.modules_to_test:
             self.assertEqual(module.is_aparam_nall(), self.expected_aparam_nall)
 
+    def test_model_output_type(self):
+        """Test model_output_type."""
+        for module in self.modules_to_test:
+            self.assertEqual(
+                module.model_output_type(), self.expected_model_output_type
+            )
+
     def test_get_nnei(self):
         """Test get_nnei."""
         expected_nnei = sum(self.expected_sel)
@@ -77,7 +83,7 @@ class AtomicModelTestCase(CommonTestCase):
             self.assertEqual(module.get_ntypes(), len(self.expected_type_map))
 
     def test_forward(self):
-        """Test forward."""
+        """Test forward and forward_lower."""
         nf = 1
         nloc = 3
         coord = np.array(
@@ -98,11 +104,26 @@ class AtomicModelTestCase(CommonTestCase):
             mixed_types=True,
             box=cell,
         )
+        ret = []
         ret_lower = []
         for module in self.modules_to_test:
-            module = forward_wrapper(module)
+            module = self.forward_wrapper(module)
+            ret.append(module(coord, atype, cell))
 
-            ret_lower.append(module(coord_ext, atype_ext, nlist))
+            ret_lower.append(module.forward_lower(coord_ext, atype_ext, nlist))
+        for kk in ret[0].keys():
+            subret = []
+            for rr in ret:
+                if rr is not None:
+                    subret.append(rr[kk])
+            if len(subret):
+                for ii, rr in enumerate(subret[1:]):
+                    if subret[0] is None:
+                        assert rr is None
+                    else:
+                        np.testing.assert_allclose(
+                            subret[0], rr, err_msg=f"compare {kk} between 0 and {ii}"
+                        )
         for kk in ret_lower[0].keys():
             subret = []
             for rr in ret_lower:
@@ -110,6 +131,25 @@ class AtomicModelTestCase(CommonTestCase):
                     subret.append(rr[kk])
             if len(subret):
                 for ii, rr in enumerate(subret[1:]):
-                    np.testing.assert_allclose(
-                        subret[0], rr, err_msg=f"compare {kk} between 0 and {ii}"
-                    )
+                    if subret[0] is None:
+                        assert rr is None
+                    else:
+                        np.testing.assert_allclose(
+                            subret[0], rr, err_msg=f"compare {kk} between 0 and {ii}"
+                        )
+        same_keys = set(ret[0].keys()) & set(ret_lower[0].keys())
+        self.assertTrue(same_keys)
+        for key in same_keys:
+            for rr in ret:
+                if rr[key] is not None:
+                    rr1 = rr[key]
+                    break
+            else:
+                continue
+            for rr in ret_lower:
+                if rr[key] is not None:
+                    rr2 = rr[key]
+                    break
+            else:
+                continue
+            np.testing.assert_allclose(rr1, rr2)
