+++ source/tests/universal/pt/atomc_model/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
